выделить подготовку данных в отдельный модуль
